+++ src/ds/app/event.cpp
@@ -22,7 +22,7 @@ Event::Event()
 {
 }
 
-Event::Event(const int what)
+Event::Event(const size_t what)
 	: mWhat(what)
 	, mEventOrigin(0.0f, 0.0f, 0.0f)
 	, mUserStringData("")
