+++ src/ds/query/recycle_array.h
@@ -30,7 +30,7 @@ public:
 public:
 	// Clients can optionally supply a factory that will initialize
 	// each element, although that only applies if T is a pointer.
-	RecycleArray(const int growBy = 1);
+	RecycleArray(const size_t growBy = 1);
 	RecycleArray(const RecycleArray &rhs);
 	virtual ~RecycleArray();
 
@@ -38,24 +38,24 @@ public:
 	const T*					data() const;
 
 	int							size() const;
-	bool						setSize(const int newSize);
-	bool						setSize(const int newSize, const T& initialize);
+	bool						setSize(const size_t newSize);
+	bool						setSize(const size_t newSize, const T& initialize);
 	void						clear();
 
-	T&							operator[](const int index) const;
+	T&							operator[](const size_t index) const;
 
 	bool						setTo(const RecycleArray<T>&);
 
-	int							find(const T&, const int startIndex = 0) const;
+	int							find(const T&, const size_t startIndex = 0) const;
 
 	bool						add(const T&);
 	// Note:  This can cause a reorder!
 	bool						remove(const T&);
-	bool						removeAt(const int index);
+	bool						removeAt(const size_t index);
 
 	// Treat like a list, and keep things ordered
 	T							popFront();
-	T							popAt(const int index);
+	T							popAt(const size_t index);
 
 	// Exchange data between the arrays
 	void						swap(RecycleArray<T>&);
@@ -63,18 +63,18 @@ public:
 	RecycleArray<T>&			operator=(const RecycleArray<T>&);
 	bool						operator==(const RecycleArray<T>&) const;
 
-	int							alloc() const;
+	size_t						alloc() const;
 
 private:
-	const int					mGrowBy;
+	const size_t				mGrowBy;
 	T*							mD;
-	int							mSize, mAlloc;
+	size_t						mSize, mAlloc;
 };
 
 /* RECYCLE-ARRAY IMPLEMENTATION
 	******************************************************************/
 template <class T>
-RecycleArray<T>::RecycleArray(const int growBy)
+RecycleArray<T>::RecycleArray(const size_t growBy)
 	: mGrowBy(growBy)
 	, mD(nullptr)
 	, mSize(0)
@@ -117,14 +117,14 @@ inline int RecycleArray<T>::size() const
 }
 
 template <class T>
-bool RecycleArray<T>::setSize(const int newSize)
+bool RecycleArray<T>::setSize(const size_t newSize)
 {
 	if(newSize <= mAlloc) {
 		mSize = newSize;
 		return true;
 	}
-	const int	newAlloc = newSize + mGrowBy;
-	T*			newD = (T*)realloc(mD, sizeof(T)*newAlloc);
+	const size_t	newAlloc = newSize + mGrowBy;
+	T*				newD = (T*)realloc(mD, sizeof(T)*newAlloc);
 	if(newD == NULL) return false;
 
 	mD = newD;
@@ -134,7 +134,7 @@ bool RecycleArray<T>::setSize(const int newSize)
 }
 
 template <class T>
-bool RecycleArray<T>::setSize(const int newSize, const T& initialize)
+bool RecycleArray<T>::setSize(const size_t newSize, const T& initialize)
 {
 	const int			oldSize = mSize;
 	if(!setSize(newSize)) return false;
@@ -143,7 +143,7 @@ bool RecycleArray<T>::setSize(const int newSize, const T& initialize)
 }
 
 template <class T>
-T& RecycleArray<T>::operator[](const int index) const
+T& RecycleArray<T>::operator[](const size_t index) const
 {
 	if(index < 0 || index >= mSize) throw bad_index();
 	return mD[index];
@@ -165,7 +165,7 @@ bool RecycleArray<T>::setTo(const RecycleArray<T>& src)
 }
 
 template <class T>
-int RecycleArray<T>::find(const T& t, const int startIndex) const
+int RecycleArray<T>::find(const T& t, const size_t startIndex) const
 {
 	for(int k = startIndex; k < mSize; k++) {
 		if(mD[k] == t) return k;
@@ -180,7 +180,7 @@ bool RecycleArray<T>::add(const T& t)
 		mD[mSize++] = t;
 		return true;
 	}
-	const int		oldSize = mSize;
+	const size_t oldSize = mSize;
 	if(!setSize(mSize + 1)) return false;
 	mD[oldSize] = t;
 	return true;
@@ -207,7 +207,7 @@ bool RecycleArray<T>::remove(const T& t)
 }
 
 template <class T>
-bool RecycleArray<T>::removeAt(const int idx)
+bool RecycleArray<T>::removeAt(const size_t idx)
 {
 	if(mSize < 1 || idx < 0 || idx >= mSize) return false;
 	mD[idx] = mD[mSize - 1];
@@ -222,7 +222,7 @@ T RecycleArray<T>::popFront()
 }
 
 template <class T>
-T RecycleArray<T>::popAt(const int index)
+T RecycleArray<T>::popAt(const size_t index)
 {
 	if(index < 0 || index >= mSize) throw bad_index();
 
@@ -235,9 +235,9 @@ T RecycleArray<T>::popAt(const int index)
 template <class T>
 void RecycleArray<T>::swap(RecycleArray<T>& o)
 {
-	T*			d = mD;			mD = o.mD;			o.mD = d;
-	const int	s = mSize;		mSize = o.mSize;	o.mSize = s;
-	const int	a = mAlloc;		mAlloc = o.mAlloc;	o.mAlloc = a;
+	T*				d = mD;			mD = o.mD;			o.mD = d;
+	const size_t	s = mSize;		mSize = o.mSize;	o.mSize = s;
+	const size_t	a = mAlloc;		mAlloc = o.mAlloc;	o.mAlloc = a;
 }
 
 template <class T>
@@ -256,7 +256,7 @@ bool RecycleArray<T>::operator==(const RecycleArray<T>& o) const
 }
 
 template <class T>
-inline int RecycleArray<T>::alloc() const
+inline size_t RecycleArray<T>::alloc() const
 {
 	return mAlloc;
 }
